refactor: move all update state functionality to a new class
