+++ port_ocean/core/event_listener/polling.py
@@ -51,7 +51,9 @@ class PollingEventListener(BaseEventListener):
         self.event_listener_config = event_listener_config
 
     def should_resync(self, last_updated_at: str) -> bool:
-        _last_updated_at = ocean.app.last_integration_state_updated_at
+        _last_updated_at = (
+            ocean.app.resync_state_updater.last_integration_state_updated_at
+        )
 
         if _last_updated_at is None:
             return self.event_listener_config.resync_on_start
@@ -78,7 +80,9 @@ class PollingEventListener(BaseEventListener):
 
             if self.should_resync(last_updated_at):
                 logger.info("Detected change in integration, resyncing")
-                ocean.app.last_integration_state_updated_at = last_updated_at
+                ocean.app.resync_state_updater.last_integration_state_updated_at = (
+                    last_updated_at
+                )
                 running_task: Task[Any] = get_event_loop().create_task(self._resync({}))
                 signal_handler.register(running_task.cancel)
