+++ port_ocean/core/handlers/resync_state_updater/__init__.py
@@ -1,0 +1,5 @@
+from .updater import ResyncStateUpdater
+
+__all__ = [
+    "ResyncStateUpdater",
+]
